+++ src/rules/import.ts
@@ -1,47 +1,46 @@
-import { truthy } from "@ezez/utils";
 import importPlugin from "@ezez/_eslint-plugin-import";
-import { fixupPluginRules } from "@ezez/__eslint__compat";
+import { truthy } from "@ezez/utils";
 
-import type { MergedOptions } from "../types.js";
 import type { Linter } from "eslint";
+import type { MergedOptions } from "../types.js";
 
-import { ERROR, WARN, OFF } from "./_states.js";
+import { ERROR, OFF, WARN } from "./_states.js";
 
 const get = (mergedOptions: MergedOptions): Linter.FlatConfig[] => {
     return [
         mergedOptions.import && {
             name: "Import - static analysis",
             plugins: {
-                // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment,@typescript-eslint/no-explicit-any,@typescript-eslint/no-unsafe-type-assertion
-                import: fixupPluginRules(importPlugin) as any,
+                import: importPlugin,
             },
             settings: {
                 "import/ignore": ["\\.(scss|saas|less|css|jpg|jpeg|gif|png|webm|webp|svg|ttf|woff|woff2)$"],
             },
             rules: {
-                "import/no-unresolved": OFF(),
-                "import/named": ERROR(),
+                "import/consistent-type-specifier-style": ERROR(),
                 "import/default": ERROR(),
+                "import/enforce-node-protocol-usage": OFF(),
+                "import/named": ERROR(),
                 "import/namespace": ERROR(),
-                "import/no-restricted-paths": OFF(),
                 "import/no-absolute-path": ERROR(),
-                "import/no-dynamic-require": OFF(),
-                "import/no-internal-modules": OFF(),
-                "import/no-webpack-loader-syntax": OFF(),
-                "import/no-self-import": ERROR(),
                 "import/no-cycle": ERROR({
                     // maxDepth: 0 - don't use until linting goes slow
                     ignoreExternal: true,
                     allowUnsafeDynamicCyclicDependency: false,
                 }),
+                "import/no-dynamic-require": OFF(),
+                "import/no-empty-named-blocks": ERROR(),
+                "import/no-internal-modules": OFF(),
+                "import/no-relative-packages": OFF(),
+                "import/no-relative-parent-imports": OFF(),
+                "import/no-restricted-paths": OFF(),
+                "import/no-self-import": ERROR(),
+                "import/no-unresolved": OFF(),
                 "import/no-useless-path-segments": ERROR({
                     noUselessIndex: false,
                     commonjs: false, // false by default, check if true will leave support for import syntax @TODO
                 }),
-                "import/no-relative-parent-imports": OFF(),
-                "import/no-relative-packages": OFF(),
-                "import/no-empty-named-blocks": ERROR(),
-                "import/consistent-type-specifier-style": ERROR(),
+                "import/no-webpack-loader-syntax": OFF(),
             },
         },
         mergedOptions.import && {
@@ -84,30 +83,7 @@ const get = (mergedOptions: MergedOptions): Linter.FlatConfig[] => {
                 }),
                 "import/no-namespace": OFF(),
                 "import/extensions": OFF(), // TODO make configurable with flat configs
-                "import/order": ERROR({
-                    "named": false,
-                    "groups": [
-                        "builtin", "external", "type", "internal", "parent", "sibling", "index", "object", "unknown",
-                    ],
-                    "pathGroups": [
-                        {
-                            pattern: "react",
-                            group: "builtin",
-                            position: "before",
-                        },
-                        {
-                            pattern: "*.{css,scss}",
-                            group: "unknown",
-                            patternOptions: { matchBase: true },
-                            position: "before",
-                        },
-                    ],
-                    "distinctGroup": true,
-                    "pathGroupsExcludedImportTypes": ["react"],
-                    "newlines-between": "always",
-                    "warnOnUnassignedImports": true,
-                    "alphabetize": undefined,
-                }),
+                "import/order": OFF(),
                 // TODO: reenable newlines-after-import after this is fixed https://github.com/import-js/eslint-plugin-import/issues/2913
                 "import/newline-after-import": OFF({
                     count: 1,