+++ src/rules/node.ts
@@ -1,7 +1,7 @@
 import type { Linter } from "eslint";
 import type { MergedOptions } from "../types.js";
 
-import { ERROR, WARN, OFF } from "./_states.js";
+import { ERROR, OFF, WARN } from "./_states.js";
 
 const get = (mergedOptions: MergedOptions): Linter.FlatConfig | null => {
     return mergedOptions.node